requirements.txt
numpy==1.21.0
pandas==1.3.3
matplotlib==3.4.3
ta-lib==0.4.0
pyyaml==6.0
jsonschema==4.4.0
seaborn==0.11.2
scipy==1.7.1
loguru==0.5.3
